pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-input-tool"
version = "0.1.0"
description = "Voice input tool using OpenAI Whisper API"
requires-python = ">=3.10"
dependencies = [
    "sounddevice>=0.4.6",
    "scipy>=1.11.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "pyperclip>=1.8.0",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.0.0",
]

[project.scripts]
voice-input = "voice_input.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/voice_input"]
